Add component files and imports/exports
