+++ src/components/Form/Form.css
@@ -1,0 +1,3 @@
+h1 {
+    color: #282c34;
+}
