+++ src/components/Form/Form.js
@@ -1,0 +1,9 @@
+import './Form.css';
+
+function Form() {
+
+    return <h1>New Form Title</h1>
+
+};
+
+export default Form;
