+++ src/components/Ideas/Ideas.css
@@ -1,0 +1,6 @@
+h2 {
+    color: #61dafb;
+    padding-right: 8px;
+    padding-left: 8px;
+    outline: solid white 1px;
+}
