+++ src/components/Ideas/Ideas.js
@@ -1,0 +1,7 @@
+import './Ideas.css';
+
+function Idea(props) {
+    return <h2>Idea Card #{props.title}</h2>
+}
+
+export default Idea
