+++ src/components/Ideas/Ideas.test.js
@@ -1,0 +1,2 @@
+import { render, screen } from '@testing-library/react';
+import Idea from './Ideas';
